% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stddiff_spark.R
\name{check_binary_group}
\alias{check_binary_group}
\title{Check if group variable has exactly 2 levels}
\usage{
check_binary_group(group_levels, gcol_name)
}
\description{
Check if group variable has exactly 2 levels
}
\keyword{internal}
