% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stddiff_spark.R
\name{validate_stddiff_inputs}
\alias{validate_stddiff_inputs}
\title{Validate inputs for stddiff functions}
\usage{
validate_stddiff_inputs(data, gcol, vcol, type)
}
\description{
Validate inputs for stddiff functions
}
\keyword{internal}
